question_generation/pyproject.toml
[project]
name = "iterate-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "bs4>=0.0.2",
    "ddgs>=9.9.1",
    "ipykernel>=7.1.0",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "keybert>=0.8.0",
    "langchain>=1.0.7",
    "langchain-anthropic>=1.0.4",
    "langchain-community>=0.4.1",
    "linkup-sdk>=0.9.0",
    "numpy>=1.24.0,<2.0.0",
    "pip>=25.3",
    "pypdf>=6.2.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "spacy>=3.8.9",
    "torch>=2.0.0,<2.3.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=12.0",
]
